separate element classes and types
