+++ moonlight-core/src/main/scala/com/github/jamedge/moonlight/core/model/neo4j/Nodes.scala
@@ -1,7 +1,7 @@
 package com.github.jamedge.moonlight.core.model.neo4j
 
 import com.github.jamedge.moonlight.core.model.IOElement
-import com.github.jamedge.moonlight.core.model.neo4j.GraphElements.{ElementClass, GraphElement, Node}
+import com.github.jamedge.moonlight.core.model.neo4j.GraphElements.Node
 
 object Nodes {
